partially completed feature: comments

+++ Models/SimplePostComment.cs
@@ -5,7 +5,7 @@ namespace isolaatti_API.Models
         public long Id { get; set; }
         public string TextContent { get; set; }
         public int WhoWrote { get; set; }
-        public int SimpleTextPostId { get; set; }
+        public long SimpleTextPostId { get; set; }
         public int Privacy { get; set; }
     }
 }
